Feature/edit question tests

+++ src/main/java/uk/gov/hmcts/reform/coh/domain/QuestionStateId.java
@@ -1,9 +1,7 @@
 package uk.gov.hmcts.reform.coh.domain;
 
 import javax.persistence.*;
-import javax.validation.constraints.NotNull;
 import java.io.Serializable;
-import java.util.Date;
 import java.util.Objects;
 
 @Embeddable
@@ -13,12 +11,12 @@ public class QuestionStateId implements Serializable {
     private Long questionId;
 
     @Column(name = "question_state_id")
-    private int questionStateId;
+    private Long questionStateId;
 
 
     public QuestionStateId(){}
 
-    public QuestionStateId(Long questionId, int questionStateId){
+    public QuestionStateId(Long questionId, Long questionStateId){
         this.questionId = questionId;
         this.questionStateId = questionStateId;
     }
@@ -50,11 +48,11 @@ public class QuestionStateId implements Serializable {
         this.questionId = questionId;
     }
 
-    public int getQuestionStateId() {
+    public Long getQuestionStateId() {
         return questionStateId;
     }
 
-    public void setQuestionStateId(int questionStateId) {
+    public void setQuestionStateId(Long questionStateId) {
         this.questionStateId = questionStateId;
     }
 }
